feat: add IntroScreen component with interactive trigram symbols and menu options
- Implemented IntroScreen with visual elements including a logo, subtitle, and trigram symbols.
- Added interactive features for selecting sparring or training modes using keyboard and mouse events.
- Integrated audio feedback for menu interactions and selections.
- Created reusable components for displaying trigrams, menu buttons, and background graphics.
- Enhanced visual effects with animations and dynamic styling based on user interactions.
