+++ tsconfig.tsbuildinfo
@@ -1,1 +1,1 @@
-{"root":["./src/App.test.tsx","./src/App.tsx","./src/main.tsx","./src/vite-env.d.ts","./src/audio/AudioManager.ts","./src/audio/AudioUtils.ts","./src/audio/DefaultSoundGenerator.ts","./src/audio/placeholder-sounds.ts","./src/audio/__tests__/AudioManager.integration.test.ts","./src/audio/__tests__/AudioManager.test.ts","./src/audio/__tests__/AudioUtils.test.ts","./src/audio/__tests__/integration.test.tsx","./src/components/__tests__/AudioIntegration.test.tsx","./src/components/__tests__/GameAudio.test.tsx","./src/components/game/DojangBackground.test.tsx","./src/components/game/DojangBackground.tsx","./src/components/game/GameEngine.test.tsx","./src/components/game/GameEngine.tsx","./src/components/game/GameUI.test.tsx","./src/components/game/GameUI.tsx","./src/components/game/HitEffectsLayer.test.tsx","./src/components/game/HitEffectsLayer.tsx","./src/components/game/Player.test.tsx","./src/components/game/Player.tsx","./src/components/game/PlayerVisuals.test.tsx","./src/components/game/PlayerVisuals.tsx","./src/hooks/useTexture.ts","./src/test/setup.ts","./src/types/pixi-react.d.ts"],"version":"5.8.3"}
+{"root":["./src/App.test.tsx","./src/App.tsx","./src/main.tsx","./src/vite-env.d.ts","./src/audio/AudioManager.ts","./src/audio/AudioUtils.ts","./src/audio/DefaultSoundGenerator.ts","./src/audio/placeholder-sounds.ts","./src/audio/__tests__/AudioManager.integration.test.ts","./src/audio/__tests__/AudioManager.test.ts","./src/audio/__tests__/AudioUtils.test.ts","./src/audio/__tests__/integration.test.tsx","./src/components/__tests__/AudioIntegration.test.tsx","./src/components/__tests__/GameAudio.test.tsx","./src/components/game/DojangBackground.test.tsx","./src/components/game/DojangBackground.tsx","./src/components/game/GameEngine.test.tsx","./src/components/game/GameEngine.tsx","./src/components/game/GameUI.test.tsx","./src/components/game/GameUI.tsx","./src/components/game/HitEffectsLayer.test.tsx","./src/components/game/HitEffectsLayer.tsx","./src/components/game/Player.test.tsx","./src/components/game/Player.tsx","./src/components/game/PlayerVisuals.test.tsx","./src/components/game/PlayerVisuals.tsx","./src/components/intro/IntroScreen.tsx","./src/hooks/useTexture.ts","./src/test/setup.ts","./src/types/pixi-react.d.ts"],"version":"5.8.3"}
